requirements.txt
flask
flask-limiter
flask-migrate
flask-restful
flask-sqlalchemy
gunicorn
python-dotenv
psycopg2
python-jose
requests
pytest
